Add `#'M:F/A` syntax as shortcut for `(fun M F A)`
In Common Lisp, `#'function-name` exists as a shortcut to the function
reference `(function function-name)`. Although that in LFE you couldn't
do that in exactly the same way, as a function's arity always has to be
known in order to reference it, we can have something very similar:

Just use the Common Lisp syntax and specify the arity in addition.

Usage Examples:

    (: lists map #'string:to_lower/1 '("fooBar" "FizzBuzz"))

    (: lists filter #'is_function/1 (list #'-/1 3 #'abs/1 -5))

    (: lists map #'-/1 '(5 -8 1))

    (: lists zipwith #'-/2 '(1 2 3) '(1 5 8))

    (: lists zipwith #'//2 '(1 2 3) '(1 5 8))

    (: lists zipwith #'=:=/2 '(1 2 3) '(1 5 8))

    (: lists sort #'</2 '(1 8 5 2 3))